==> SDMs/2_SDMs_reduce_spatial_autocorrelation.R <==
##Based on code generated by Peter Breslin and Fabio Suzart de Albuquerque (ASU)

######################################################################
########################### Libraries ################################
######################################################################

library(sp)
library(raster)
library(rgeos)

#####################################
############ Load Files #############
#####################################
butternut_drive <- "C:\\Users\\eschumacher\\Documents\\GitHub\\butternut"

##set wd
setwd(butternut_drive)

##read in occurrence records
butternut_presence <- read.csv("SDMs\\InputFiles\\occurrence_records\\occurrence_records_all_red.csv")

##remove extra oc column
butternut_presence <- butternut_presence[,-1]

##set extent 
min_lon <- min(butternut_presence$Longitude)
max_lon <- max(butternut_presence$Longitude)
min_lat <- min(butternut_presence$Latitude)
max_lat <- max(butternut_presence$Latitude)

##set spatial components of the data frame 
coordinates(butternut_presence) <- c('Longitude', 'Latitude')
proj4string(butternut_presence) <- CRS("+proj=longlat +ellps=WGS84 +datum=WGS84")

##Projection string
projection <- c("+proj=aea +lat_1=20 +lat_2=60 +lat_0=40 +lon_0=-96 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs")

##Load raster extent
extent_raster <- raster("SDMs\\InputFiles\\elevation_extent.tif")

##project Raster
extent_project <- projectRaster(extent_raster, crs = projection)

##write out 
writeRaster(extent_project, "extent_project.tif")

######################################################################
######## Project and Remove Spatial Autocorrelation ##################
######################################################################
##Project to Albers Equal Area Conic 
butternut_presence_2 <- spTransform(butternut_presence, CRS(paste0(projection)))

##First remove points at the exact same place
butternut_presence_2 <- remove.duplicates(butternut_presence_2)

##Reduce spatial autocorrelation, identifying points within certain distances
#This code will throw warnings but it is working 

#####100 meters
points_matrix <- gWithinDistance(butternut_presence, dist = .00086206895, byid = TRUE)
#####300 meters
points_matrix <- gWithinDistance(butternut_presence, dist = .00258620685, byid = TRUE)
######500 m
points_matrix <- gWithinDistance(butternut_presence, dist = .0043103448, byid = TRUE)
#####1 km
points_matrix <- gWithinDistance(butternut_presence, dist = .0086206897, byid = TRUE)
####2km
points_matrix <- gWithinDistance(butternut_presence, dist = .0172413794, byid = TRUE)

##create tri matrix with NAs for bottom rows
points_matrix[lower.tri(points_matrix, diag=TRUE)] <- NA

##now determine where points within 1 km are 
auto_cols <- colSums(points_matrix, na.rm=TRUE) == 0

##now subset the original data by this 
occurrence_wo_auto <- butternut_presence[auto_cols,]

##convert to a data frame
occurrence_no_auto <- data.frame(occurrence_wo_auto) ##reduced to 3053 individuals 

##occurrence no autocorrelation write out 
write.csv(occurrence_no_auto, "SDMs\\InputFiles\\occurrence_noauto_noproj.csv")

##Check to make sure extents align
setwd("SDMs\\OutputFiles")
pdf("extent_pointsnoauto.pdf")
plot(extent_raster)
points(occurrence_wo_auto, pch = 16, col = "dodgerblue")
dev.off()
