==> SDMs/3_SDMs_generate_pseudoabsence.R <==
##Based on code generated by Peter Breslin and Fabio Suzart de Albuquerque (ASU)

######################################################################
########################### Libraries ################################
######################################################################

library(sp)
library(raster)
library(dismo)
library(rgeos)

#####################################
############ Load Files #############
#####################################
butternut_drive <- "C:\\Users\\eschumacher\\Documents\\GitHub\\butternut"

##set wd
setwd(butternut_drive)

######load in presence records
butternut_pres <- read.csv("SDMs\\InputFiles\\occurrence_noauto_noproj.csv")

##remove extra column before longitudes and latitudes
butternut_pres <- butternut_pres[,-1]

##calculate crop limits
min_lon <- min(butternut_pres$Longitude)
max_lon <- max(butternut_pres$Longitude)
min_lat <- min(butternut_pres$Latitude)
max_lat <- max(butternut_pres$Latitude)

##spatial code 
coordinates(butternut_pres) <- c('Longitude', 'Latitude')
proj4string(butternut_pres) <- CRS("+proj=longlat +ellps=WGS84 +datum=WGS84")

##Projection string
projection <- c("+proj=aea +lat_1=20 +lat_2=60 +lat_0=40 +lon_0=-96 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs")

##project presence points
butternut_pres <- spTransform(butternut_pres, projection)

##write in extent raster
extent_project <- raster("SDMs\\InputFiles\\extent_project.tif")

######################################################################
################## Generate Pseudo-Absence Points ####################
######################################################################
##create data frame of lon/lat
butternut_pres_df <- data.frame(butternut_pres)

##create a document for points = presence records 
pres_length <- length(butternut_pres_df[,1])

##Now start generating background points
background_points <- randomPoints(extent_project, pres_length)

##now continue to sample
set.seed(25)
background_2 <- gridSample(background_points, extent_project, n=1) 
bg_proj <- SpatialPoints(background_2, proj4string=CRS(projection))

##now set occurrence records to a spatial gridded object
set.seed(8)
extent_sample <- gridSample(butternut_pres, extent_project, n=1) 

##turn into a spatial points object
occurrence_grid_spatial <-SpatialPoints(extent_sample, CRS(projection))

##determine the difference between absence and presence to make sure they aren't in the same place 
butternut_abs <- gDifference(bg_proj, occurrence_grid_spatial)

##extract to points to make sure data is clean
abs_ext <- extract(extent_project, butternut_abs)
abs_ext_df <- data.frame(abs_ext)
abs_ext_df <- na.omit(abs_ext_df) ##no points get removed

##no na points, good to go
##Butternut absence data frame
butternut_abs_df <- data.frame(butternut_abs)
colnames(butternut_abs_df) <- c("Absence_points")

##write out absence 
write.csv(butternut_abs_df, "SDMs\\InputFiles\\butternut_abs.csv")

##Now rename columns and set up data frame
col_names <- colnames(butternut_pres_df)

colnames(butternut_abs_df) <- col_names

##Prepare presence and absence records
##Absence column added 
butternut_abs_df$PA <- "0"

##Add presence column to presence points
butternut_pres_df$PA <- "1"

##combine into one data frame
butternut_pa <- rbind(butternut_pres_df, butternut_abs_df)

##Write out files 
write.csv(butternut_pa, "SDMs\\InputFiles\\butternut_pa.csv")

##Visualize results 
pdf("SDMs\\OutputFiles\\pres_abs_points.pdf", width = 6, height = 6)
plot(extent_project)
points(butternut_abs, pch = 16, col = "dodgerblue4")
points(butternut_pres, pch = 16, col = "dodgerblue")
legend('topright', legend = c("Presence", "Absence"), pch = 16, col = c("dodgerblue", "dodgerblue4"))
dev.off()
