==> SDMs/4_SDMs_variable_selection.R <==
###Based on code generated by Peter Breslin (ASU) and Fabio Suzart de Albuquerque 
######################################################################
########################### Libraries ################################
######################################################################

library(sp)
library(raster)
library(tidyverse)
library(dendextend)
library(ltm)

#####################################
############ Load Files #############
#####################################
butternut_drive <- "C:\\Users\\eschumacher\\Documents\\GitHub\\butternut"

##set wd
setwd(butternut_drive)

##load in presence/absence points file
butternut_pa <- read.csv("SDMs\\InputFiles\\butternut_pa.csv")

##remove unnecessary column
butternut_pa <- butternut_pa[,-1]

##load in elevation crop
elevation_crop <- raster("SDMs\\InputFiles\\elevation_extent.tif")

##load in projected extent
extent_project <- raster("SDMs\\InputFiles\\extent_project.tif")

##Projection string
projection <- c("+proj=aea +lat_1=20 +lat_2=60 +lat_0=40 +lon_0=-96 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs")

####################################
##### Variable Selection ###########
####################################
########Load worldclim documents
setwd("SDMs\\InputFiles\\bio_2-5m_bil")

##create multiple lists
worldclim_list = list.files(pattern = ".bil$")
worldclim_raster_list <- list()
worldclim_crop_list <- list()
worldclim_project_list <- list()

##loop through soil docs to crop and project every layer 
for(j in 1:length(worldclim_list)) {
  #convert to rasters
  worldclim_raster_list[[j]] = raster(paste0(butternut_drive,"\\SDMs\\InputFiles\\bio_2-5m_bil\\", worldclim_list[[j]]))
  
   worldclim_crop_list[[j]] <- crop(worldclim_raster_list[[j]], elevation_crop)
  
   worldclim_project_list[[j]] <- projectRaster(worldclim_crop_list[[j]], crs = projection)
  
}

##layer names 
worldclim_names <- c("MAT","MTwarQ","MTCQ","MAP","PwetM","PDM","precip_season","PwetQ","PDQ","PwarQ","PCQ",
                     "MDR","ISO","temp_season","MTwarM","MTCM","temp_range","MTwetQ","MTDQ")

##Rename layers
for(p in 1:length(worldclim_names)) {
  
  names(worldclim_project_list[[p]]) <- worldclim_names[[p]]
  
}

##now create a raster stack of these variables
worldclim_stack <- stack(worldclim_project_list[[1]], worldclim_project_list[[2]],
                         worldclim_project_list[[3]], worldclim_project_list[[4]],
                         worldclim_project_list[[5]], worldclim_project_list[[6]],
                         worldclim_project_list[[7]], worldclim_project_list[[8]],
                         worldclim_project_list[[9]], worldclim_project_list[[10]],
                         worldclim_project_list[[11]], worldclim_project_list[[12]],
                         worldclim_project_list[[13]], worldclim_project_list[[14]],
                         worldclim_project_list[[15]], worldclim_project_list[[16]],
                         worldclim_project_list[[17]], worldclim_project_list[[18]],
                         worldclim_project_list[[19]])

##save stack in input files 
setwd(paste0(butternut_drive, "\\SDMs\\OutputFiles"))
save(worldclim_stack, file = "worldclim_stack.RData")

##now extract values
worldclim_extract <- extract(worldclim_stack, butternut_pa[,1:2])
worldclim_df <- data.frame(worldclim_extract)

##bind into columns 
butternut_var <- cbind(butternut_pa, worldclim_df)

##remove nas
butternut_var <- na.omit(butternut_var)

##rename
colnames(butternut_var) <- c("Longitude", "Latitude", "PA",  
                             names(worldclim_stack))
                             
                             
##check out how many individuals are left 
length(butternut_var[butternut_var$PA == "1",]$PA) ##Presence points, 3051
length(butternut_var[butternut_var$PA == "0",]$PA) ##Absence points, 3048

##write out points with variables
write.csv(butternut_var, paste0(butternut_drive,"\\SDMs\\InputFiles\\butternut_var.csv"))

####################################
####### Variable Correlation #######
####################################
######correlation matrix
predictors_correlation <- cor(butternut_var[,4:22])

#correlation matrix as distance matrix - dissimilarity matrix
predictors_dist <- as.dist(abs(predictors_correlation))

#create dissimilarity tree 
predictor_cluster <- hclust(1-(predictors_dist))

##now create dendrogram 
predictor_dendrogram <- as.dendrogram(predictor_cluster)

##set names 
worldclim_full_names <- c('Mean Temperature of the Warmest Quarter (C)',"Maximum Temperature of the Warmest Month (C)", 
                          "Mean Temperature of the Driest Quarter(C)","Isothermality","Mean Annual Temperature (C)",
                          "Mean Temperature of the Coldest Quarter (C)", "Mean Temperature of the Coldest Month (C)",
                          "Mean Diurnal Range (C)","Precipitation of the Warmest Quarter (mm)",
                          "Precipitation of the Wettest Month (mm)", "Precipitation of the Wettest Quarter (mm)",
                          "Mean Temperature of the Wettest Quarter (C)", "Seasonal Precipitation (mm)",
                          "Temperature Seasonality (C)", "Annual Temperature Range (C)",
                          "Mean Annual Precipitation (mm)", "Precipitation of the Coldest Quarter (mm)",
                          "Precipitation of the Driest Month (mm)", "Precitation of the Driest Quarter (mm)"
                          )

##make labels pretty
predict_dend_2 <- predictor_dendrogram %>% 
                    set("labels", worldclim_full_names) %>% 
                    set("labels_cex", 0.8)
                  
##create dendrogram figure
pdf(paste0(butternut_drive, "\\SDMs\\OutputFiles\\ecogeo_dendrogram.pdf"), width = 10, height = 6)
par(mar = c(2,2,2,20))

predict_dend_2 %>% 

  plot(main="Ecogeographic Dissimilarity Dendrogram", horiz = TRUE) %>% 
  
  abline(v= 0.5, col = "red")

dev.off()

#Biserial correlation
variable_names <- names(butternut_var[,1:22])

##create matrix dimensions
cor_result <- length(variable_names)
butternut_col <- 2 

##create matrix to store biserial correlation
biserial_cor_matrix <- matrix(ncol=butternut_col, nrow=cor_result)

##calculate biserial correlation and store in matrix
for (i in 4:length(variable_names)){
  
  biserial_cor_coef <- biserial.cor(butternut_var[,i],butternut_var$PA)
  
  biserial_cor_matrix[i,] <- rbind(variable_names[i], biserial_cor_coef)
  
}

##update output 
biserial_cor_out <- data.frame(biserial_cor_matrix[-1:-3,])
colnames(biserial_cor_out) <- c("Variable", "Correlation Coef")

##write out data frame
write.csv(biserial_cor_out, paste0(butternut_drive, "\\SDMs\\OutputFiles\\biserial_cor_matrix.csv"))

####write out rasters
writeRaster(worldclim_stack, paste0(butternut_drive, "\\SDMs\\OutputFiles\\worldclim_stack.tiff"))
