==> SDMs/5_SDMs_BRT_predictionmaps.R <==
##Based on code generated by Peter Breslin (ASU) and Fabio Suzart de Albuquerque 

######################################################################
########################### Libraries ################################
######################################################################

library(sp)
library(raster)
library(gbm)
library(AUC)
library(dismo)
library(ggplot2)

#####################################
############ Load Files #############
#####################################
butternut_drive <- "C:\\Users\\eschumacher\\Documents\\GitHub\\butternut"

##set wd
setwd(butternut_drive)

##load in variable data frame
butternut_var <- read.csv("SDMs\\InputFiles\\butternut_var.csv")

##Remove extra column
butternut_var <- butternut_var[,-1]

##load in elevation crop
elevation_extent <- raster("SDMs\\InputFiles\\elevation_extent.tif")

##load in extent projection 
extent_project <- raster("SDMs\\InputFiles\\extent_project.tif")

##load in raster stack
load("SDMs\\OutputFiles\\worldclim_stack.RData")

#####Projection
projection <- c("+proj=aea +lat_1=20 +lat_2=60 +lat_0=40 +lon_0=-96 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs")

######################################################################
####################### Running the Model ############################
######################################################################
##now limit to important variables
#butternut_list_variables <- c("PwetM", "MDR","MTDQ", "MTwetQ", "precip_season")
butternut_list_variables <- c("PwetM", "MDR","MTDQ", "MTwetQ", "precip_season")

##now load variables for model running
butternut_sel_var <- butternut_var[,butternut_list_variables]
butternut_sel_df <- cbind(butternut_var[,1:3], butternut_sel_var)
butternut_sel_df[,3] <- as.numeric(butternut_sel_df[,3]) 

##now run models
par(mfrow=c(1,1))
set.seed(10)
butternut_samp <- sample(nrow(butternut_sel_df), round(0.70 * nrow(butternut_sel_df))) 
testdata <- na.omit(butternut_sel_df[butternut_samp,])
traindata <- na.omit(butternut_sel_df[-butternut_samp,])

#Model calibration. Using multiple BRT parameters to identify the best number of trees. See the Elith tutorial on what these parameters mean. 
###there's a bunch of other things but code line 384 is how you do the predictions and generate the map. Most of the other things in here 
####are either in the Elith tutorial or were related to finding range contraction and 
####expansion. 
bf=c(.7)
lr=c(0.007)
tc=c(2)
ss=c(25)
#Initiate the loop to identify the best number of trees. Call your model whatever you want- the "gbm.x = 4:7" are the columns of your climate variables and the "gbm.y=3" is the column on your csv file that has 0's and 1's for presence/absence 
for (i in 1:length(lr)){
  for (j in 1:length(bf)){
    for (k in 1:length(tc)){
      for(l in 1:length(ss)){
        
        
        ##Fit the gbm model
        #create a model name to store in the results table, I removed for (l in 1:length(ss)), ss[l],{ dist[t],'_'and 
        butternut_model_name <- paste('butternut_model_',tc[k],'_',lr[i],'_',bf[j],'_',ss[l],'_',sep='')
        
        set.seed(10)
        butternut_model <- gbm.step(traindata, gbm.x = 4:length(butternut_sel_df), gbm.y = 3, family = "bernoulli", step.size = ss[l],tree.complexity = tc[k],learning.rate = lr[i], bag.fraction = bf[j])
        
        #Model internal metrics 
        cv.AUC = butternut_model$cv.statistics$discrimination.mean 
        cv.dev = butternut_model$cv.statistics$deviance.mean # average residual deviance over all folds during cross-validation 
        
        #Calculate the % of deviance explained. Deviance explained corresponds to the percentage of deviance for the null model explained by the fitted model (dev = 100% for a perfect model)
        #The function calc.deviance() in dismo package calculates the deviance between 2 vectors v 1 and v 2
        #calculating deviance explained by the model over the training dataset (int.dev)
        int.null.deviance = butternut_model$self.statistics$mean.null 
        int.residual.deviance = butternut_model$cv.statistics$deviance.mean 
        int.dev=(int.null.deviance-int.residual.deviance)/int.null.deviance
        pred.train = predict.gbm(butternut_model, traindata, n.trees = butternut_model$gbm.call$best.trees, type = "response") 
        train=calc.deviance(traindata$PA, pred.train)
        
        ##Model external metrics , calculated over an evaluation dataset evaluation
        pred_test = predict.gbm(butternut_model, testdata, n.trees = butternut_model$gbm.call$best.trees, type = "response") 
        d <- cbind(testdata$PA, pred_test) 
        pres <- d[d[,1]==1, 2]
        abs <- d[d[,1]==0, 2]
        eval <- evaluate(p=pres, a=abs)
        ext.AUC <- eval@auc
        
        #estimating the threshold to convert suitability values into binary maps (0/1)
        th<-eval@t[which.max(eval@TPR+eval@TNR)]
        
        #independent=calc.deviance(testdata$PA, pred_test)  
        #residual deviance was calculated between observed presence/absence data (v 1) and predicted probability of presence (v 2 ),
        ext.residual.deviance = calc.deviance(testdata$PA, pred_test, calc.mean=T)
        ext.null.deviance = calc.deviance(testdata$PA,rep(mean(testdata$PA),nrow(testdata)), calc.mean=T) 
        #Calculating deviance explained in the evaluation dataset. Portion of data withheld during model building
        ext.dev=(ext.null.deviance - ext.residual.deviance)/ext.null.deviance
        
        param_df <- rbind(data.frame(butternut_model_name,tc[k],lr[i],bf[j],ss[l],butternut_model$gbm.call$best.trees,cv.AUC,ext.AUC,int.dev,ext.dev,train,independent,ext.residual.deviance))
        
      }
    }
  }
}
th

##check red text for tree number 
tree_number <- 3900

##summarize model - percent contribution of each variable
summary(butternut_model)

###create list of variables in order with % contribution of each variable
butternut_predictors <- c("Seasonal Precipitation (mm)", "Mean Temperature of the Driest Quarter (C)",
                        "Mean Temperature of the Wettest Quarter (C)", "Precipitation of the Wettest Month (mm)", 
                        "Mean Diurnal Range")
##create factors
butternut_predictors <- factor(butternut_predictors, levels= c("Seasonal Precipitation (mm)", 
                                                               "Mean Temperature of the Driest Quarter (C)",
                                                              "Mean Temperature of the Wettest Quarter (C)", 
                                                              "Precipitation of the Wettest Month (mm)",
                                                              "Mean Diurnal Range"))

##data frame combination with the % contribution and the factors
butternut_influences <- data.frame(butternut_predictors,percentages=c(28.9, 28.6, 20.6, 11.5, 10.4))

##create barplot with % contribution of each variable
butternut_contribution_bp <- ggplot(data=butternut_influences, aes(x=butternut_predictors, y=percentages)) +
  geom_bar(stat="identity", color="black", fill="dodgerblue",width=.4)+
  geom_text(aes(y=percentages, label=paste0(percentages, '%')), vjust=.5, hjust=-.1,size=3)+
  theme_minimal()+ggtitle("Relative Influence of Predictors (CV AUC = 0.824)")+theme_light(base_size=14)+theme(plot.margin = unit(c(1,1,1,1), "cm"))+ylim(0,100)+
  xlab("Predictors") + ylab("Percent Contribution to the Model")

##flip it to be horizontal
butternut_contribution_bp2 <- butternut_contribution_bp + coord_flip()

setwd(paste0(butternut_drive,"\\SDMs\\OutputFiles"))
##write out file 
pdf("contribution_bp_allvar.pdf", width = 10, height = 8)
butternut_contribution_bp2
dev.off()  

################################
######## Prediction map ########
################################
########Now begin to stack variables that contribute most to the model
##to get list of names for ordering stack 
var_list <- butternut_model$var.names

##now stack
butternut_model_stack <- stack(worldclim_stack$PwetM, worldclim_stack$MDR,
                               worldclim_stack$MTDQ, worldclim_stack$MTwetQ,
                               worldclim_stack$precip_season)
  
  
##Names must be equal to compare 
names(butternut_model_stack) <- butternut_model$var.names

##now create prediction map
butternut_prediction_map <- predict(butternut_model_stack, butternut_model,n.trees=tree_number, type='response')

##write out Raster
writeRaster(butternut_prediction_map, paste0(butternut_drive, "\\SDMs\\OutputFiles\\hsm_worldclim_only_raster.tif"),
            overwrite=TRUE)

#############################
######## Plot maps ##########
#############################
setwd(butternut_drive)

##Load in presence and absence points 
butternut_pres <- butternut_sel_df[butternut_sel_df$PA == 1,]
butternut_abs <- butternut_sel_df[butternut_sel_df$PA == 0,]

##create points 
coordinates(butternut_pres) <- c('Longitude', 'Latitude')
proj4string(butternut_pres) <- CRS(projection)

coordinates(butternut_abs) <- c('Longitude', 'Latitude')
proj4string(butternut_abs) <- CRS(projection)

##now look at plots
pdf("SDMs\\OutputFiles\\hsm.pdf", width = 6, height = 6)
plot(butternut_prediction_map, main = "Habitat Suitability Model")
dev.off()

pdf("SDMs\\OutputFiles\\hsm_p.pdf", width = 6, height = 6)
plot(butternut_prediction_map, main = "HSM Presence")
points(butternut_pres, col = "dodgerblue", pch = 16)
dev.off()

pdf("SDMs\\OutputFiles\\hsm_a.pdf", width = 6, height = 6)
plot(butternut_prediction_map, main = "HSM Absence")
points(butternut_abs, col = "dodgerblue4", pch = 16)
dev.off()

pdf("SDMs\\OutputFiles\\hsm_pa.pdf", width = 6, height = 6)
plot(butternut_prediction_map, main = "HSM Presence and Absence")
points(butternut_abs, col = "dodgerblue4", pch = 16)
points(butternut_pres, col = "dodgerblue", pch = 16)
legend('topright', legend = c("Presence", "Absence"), pch = 16, col = c("dodgerblue", "dodgerblue4"))
dev.off()

######################################################################
############################# Hindcast SDMs ##########################
######################################################################
##Load in layers 
setwd("SDMs\\InputFiles\\Paleo_Files")

##time periods list
time_periods <- list.files(pattern = "_v1_2_5m")

##variables list 
var_list <- c("bio_13.tif", "bio_2.tif","bio_9.tif","bio_8.tif",
              "bio_15.tif")

##raster list
paleo_raster <- list(list(), list(), list(), list(), list(),
                     list(), list(), list())

##create paleo stacks
paleo_stack <- list()

##crop paleo rasters
paleo_crop <- list()

##project them into albers
paleo_project <- list()

##list for habitat suitability models 
hsm_list <- list()

###Write a loop to load in past raster files 
for(j in 1:length(time_periods)){
  for(k in 1:length(var_list)){
  paleo_raster[[j]][[k]] <- raster(paste0(time_periods[[j]], "\\", var_list[[k]]))
  
  ##stack rasters
  paleo_stack[[j]] <- stack(paleo_raster[[j]])
  
  ##crop them
   paleo_crop[[j]] <- crop(paleo_stack[[j]], elevation_extent)
  
  ##project the crop
   paleo_project[[j]] <- projectRaster(paleo_crop[[j]], crs = projection)
  
  }
  ##name the variables 
  names(paleo_project[[j]]) <- butternut_model$var.names
}


###Create predicted suitable habitat models for all time periods
setwd(paste0(butternut_drive,"\\SDMs\\OutputFiles\\PaleoClim_HSMs"))

##create names for time period 
time_period_names <- gsub("_.*","",time_periods)

##years before present for each time period 
ybp <- c("14.7-12.9 ka YBP", "11.7-8.326 ka YBP", "17.0-14.7 ka YBP", "21 ka YBP", "4.2-0.3 ka YBP", "130 ka YBP",
         "8.326-4.2 ka YBP", "12.9-11.7 ka YBP")

#generate habitat suitability models for all time periods 
for(i in 1:length(paleo_project)){
  
  ##use predict to calculate habitat suitabiltiy from current model
  hsm_list[[i]] <- predict(paleo_project[[i]],butternut_model,n.trees=tree_number,type='response')

  ##write Rasters out
  #writeRaster(hsm_list[[i]], paste0(time_period_names[[i]], "_hsm.tif"), overwrite = TRUE)
  
  #pdf(paste0(time_period_names[[i]], "_(", ybp[[i]], ")_hsm.pdf"))
  #plot(hsm_list[[i]], main = ybp[[i]])
  #dev.off()
  
}
